Add Endpoint for Login
This change resolves #3

It modifies the `.env` file expectations (for `configuration.py`) to remove any TP-Link Kasa credentials / information and instead has information about the API user encryption.

This change introduces the concept of Fast API routers, and sets one up for `users`.

This change also introduces the concept of a `dependencies.py` which is documented here: https://fastapi.tiangolo.com/tutorial/bigger-applications/#dependencies

This change also introduces Fast API dependency injection, documented here: https://fastapi.tiangolo.com/tutorial/dependencies/

This change introduces a `UserDatabase` which is nothing more than an in-memory cache of users that have successfully logged in to the TP-Link API service. This allows requests that require that login to verify that the user making the request has, in fact, logged in before.

The `TPLinkService` can no longer be constructed with credentials used for login, but instead will require that a new `login` method be called to get an API access token (kept in memory)

> Note: In the future, the auth token provided by the TP-Link API service on login will need to be kept along with `User` data to allow for multiple users. Ideally, the `TPLinkDeviceManager` would provide us with an official way to get that token rather than our current (unofficial) method of accessing an internal property directly

> Note that the TPLinkService instance created for this API will use whatever credentials were last provided to it via the OAuth 2.0 Login workflow

> Note that we intentionally are not using `current_user` data for "power data" endpoints at this time, but this will change in the future to support multiple users

This change also updates existing dependencies as there were some initial conflicts when running it on Python 3.9 with the new dependencies