Adding retries to test requests
`timeout` doesn't function as intended and apparently adding retries simply using the `HTTPAdapter` doesn't add any delay at all.

So, I've introduced a new dependency, `urllib3`, which will allow for retries with backoff. We probably won't need to add jitter since these are just tests but it's something to consider in the future.