Add Multi-User Support to Power Endpoints
This change leverages the auth token provided by the caller to make requests to the TP-Link API.

This cleans out _almost all_ of the previous authentication code as it was not needed given we are getting a token from a third party service.

Note that in addition to removing dependencies on the following:
* `python-jose[cryptography]==3.3.0`
* `passlib[bcrypt]==1.7.4`

We are updating the `tplink-cloud-api` dependency to `3.0.0` which has support for separate login and auth_token functionality.

> Note that the `configuration.py` is no longer being used for anything, but I am intentionally leaving it there and imported in the `dependencies.py` so it is clear how we would use it in the future when or if it becomes necessary - we may later decide to tear it out completely