Removing API Code
The API code has been migrated to this repo: https://github.com/piekstra/tplinkcloud-service

That repository will be publishing container images of the API to GitHub's Container Registry for use by this project.

This initial change is removing the migrated API code, and a follow-up change will flatten the code out from the `ui` folder to the root level of the repository.

* The `api` folder and everything within it is being deleted
* The GitHub Actions Workflow responsible for building, running, and testing the API code is being deleted
* the `docker-compose` is being updated to refer to the current latest published version of the api container image (and should only be updated as new API features are published and desired)
* The `README` has some API-specific commentary removed